Output formatters are now separate modules
